Casting some time_t's to unsigned long's to get rid of warnings.

+++ src/alert.c
@@ -836,8 +836,8 @@ int alert_expire(void) {
                 fprintf(stderr,
                     "alert_expire: Clearing %s, current: %lu, alert: %lu\n",
                     temp->unique_string,
-                    time(NULL),
-                    temp->expiration);
+                    (unsigned long)time(NULL),
+                    (unsigned long)temp->expiration);
             }
 
             // Free the storage space
@@ -925,8 +925,8 @@ clear_dangerous();
         if (debug_level & 2) {
             fprintf(stderr,
                 "Newest Alert Expired->Clearing, current: %lu, alert: %lu\n",
-                time(NULL),
-                entry->expiration );
+                (unsigned long)time(NULL),
+                (unsigned long)entry->expiration );
         }
 
 clear_dangerous();
@@ -1027,7 +1027,7 @@ return(entry);
         fprintf(stderr,"%s %s %lu\n",
             entry->to,
             entry->title,
-            entry->expiration);
+            (unsigned long)entry->expiration);
     }
 
 clear_dangerous();
